backend/pyproject.toml
[project]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = [{ name = "Seu nome aqui", email = "seu.email@email.com.br" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6,<6.0.0",
    "djangorestframework>=3.15.1,<4.0.0",
    "celery>=5.3.6,<6.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "redis>=5.0.4,<6.0.0",
    "drf-yasg>=1.21.7",
    "django-celery-beat>=2.8.1",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.7.0",
    "python-dotenv>=1.1.0",
]


[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-django>=4.11.1",
]
